// src/company/types.ts
export type CompanyType = {
  id: number;
  organization: string;
  products: string[];
  marketValue: string;
  address: string;
  ceo: string;
  country: string;
  noOfEmployees: number;
  employees: string[];
  createdAt: string;
  updatedAt: string;
};

export type CompanyModel = {
  id: number;
  organization: string;
  products: string[];
  marketValue: string;
  address: string;
  ceo: string;
  country: string;
  noOfEmployees: number;
  employees: string[];
};